Rename module to plugin name
