+++ webviz_subsurface/plugins/__init__.py
@@ -50,4 +50,4 @@ from ._surface_with_grid_cross_section import SurfaceWithGridCrossSection
 from ._surface_with_seismic_cross_section import SurfaceWithSeismicCrossSection
 from ._well_cross_section import WellCrossSection
 from ._well_cross_section_fmu import WellCrossSectionFMU
-from ._ahm_analysis_plugin import AssistedHistoryMatchingAnalysis
+from ._assisted_history_matching_analysis import AssistedHistoryMatchingAnalysis
